requirements.txt
tensorflow
keras
pandas
opencv-python
Pillow
matplotlib
scikit-learn
Flask
Pillow
